react コンポーネント化 #2
